// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck
/*
 * This file exports all enum related types from the schema.
 *
 * 🟢 You can import this file directly.
 */

export const UserRole = {
  ADMIN: "ADMIN",
  USER: "USER",
} as const;

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserStatus = {
  BLOCKED: "BLOCKED",
  ACTIVE: "ACTIVE",
  VERIFIED: "VERIFIED",
} as const;

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const offer_types = {
  INTERNET: "INTERNET",
  MINUTE: "MINUTE",
  BUNDLE: "BUNDLE",
  CALL_RATE: "CALL_RATE",
} as const;

export type offer_types = (typeof offer_types)[keyof typeof offer_types];

export const sim_type = {
  PRE_PAID: "PRE_PAID",
  POST_PAID: "POST_PAID",
} as const;

export type sim_type = (typeof sim_type)[keyof typeof sim_type];

export const network_type = {
  GRAMEENPHONE: "GRAMEENPHONE",
  ROBI: "ROBI",
  AIRTEL: "AIRTEL",
  BANGLALINK: "BANGLALINK",
  TELETALK: "TELETALK",
  SKITTO: "SKITTO",
} as const;

export type network_type = (typeof network_type)[keyof typeof network_type];

export const banks = {
  BKASH: "BKASH",
  NAGAD: "NAGAD",
  ROCKET: "ROCKET",
  UPAY: "UPAY",
  NA: "NA",
} as const;

export type banks = (typeof banks)[keyof typeof banks];